Split the path for accepting a sharing
When a user has received an email for a sharing, they can go to the
preview page. From here, there are two possibilities for accepting the
sharing: just creating a shortcut, or the full synchronization.

Before this commit, the user was first asked their Cozy URL, and on the
OAuth authorize page, they have a checkbox for the synchronize option.
Now, the choice is made on the preview page (Drive will add a
`shortcut=true` parameter in the query string on the
`GET /sharings/:id/discovery` endpoint if needed). And if it is just a
shortcut, we will skip the OAuth authorize page.