Fix expires_at: null in CouchDB
This issue was discovered by @Merkur39. When creating an
io.cozy.permissions for a share by link, the expires_at field was set to
null in CouchDB, even if the field was declared as omitempty. A nil
pointer to a time.Time is not considered by Go as an empty value for a
field of the type interface{}. So, we need to use the empty value of
type interface{} instead to fix the JSON marshaling.