fix: OIDC username with dot
We can have UserInfoField with "dotted string". In that case,
cloudery remove the dot to create a FQDN without it.

So with username "toto.tata", cloudery will create
"tototata.mycozycloud.com".

But then, when the stack does the OIDC check, stack
doesn't remove the dot from the username. So we have
inst.domain === tototata.mycozy.cloud but for the
constructed fqdn we have
toto.tata.mycozy.cloud.

result: error 500, Invalid domains

I don't know if this is the best way to fix the issue.