Preserve RAM at RAG indexing by streaming files
At RAG indexation time, we used to read the file from disk, keep it in
memory in a buffer and send it to the network.  This can be problematic
if we have plenty of parallel indexing jobs, as the RAM consumption
could skyrocket.  So, we now stream the file in a Pipe by chunks of 32KB
(the default size of `io.Copy`), avoiding RAM peak consumption even with
large files.