feat: 409 on conflict for nextcloud
The idea is to have the same behavior on the downstream route as we have
on file routes in general. If a file with the same name already exists,
then we return a 409.

To avoid introducing a breaking change, I added a parameter so the
existing behavior doesn’t change.